Enable linting on the fly
Note that this represents a slight loss of functionality. With the
update to the new API the file was explicitly specified, which according
to the documentation enables checking for "fatal errors (like undefined
functions)".

Since we are linting on the fly we can't specify a file and thus lose
this functionality.